pyproject.toml
[tool.poetry]
name = "character-ai-bot"
version = "0.1.0"
description = "Character AI bot."
authors = ["Dmitrii Kopeikin <kopeikindp@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "character_ai_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.0.0"
redis = "^5.0.0"
alembic = "^1.12.0"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.7"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
sqlalchemy-utils = "^0.41.1"
amplitude-analytics = "^1.1.3"
requests = "^2.31.0"
djlint = "^1.32.1"
jinja2 = "^3.1.2"
httpx = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79

